@@ -1,18 +1,26 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package apps_test
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"testing"
 
 	"github.com/epinio/epinio/acceptance/helpers/proc"
 	"github.com/epinio/epinio/acceptance/testenv"
-	epinioConfig "github.com/epinio/epinio/internal/cli/config"
+	"github.com/epinio/epinio/internal/cli/settings"
 
-	"github.com/onsi/ginkgo/config"
-
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
@@ -31,23 +39,23 @@ var (
 )
 
 var _ = BeforeSuite(func() {
-	fmt.Printf("Running tests on node %d\n", config.GinkgoConfig.ParallelNode)
+	fmt.Printf("Running tests on node %d\n", GinkgoParallelProcess())
 
 	testenv.SetRoot("../..")
 	testenv.SetupEnv()
 
-	nodeSuffix = fmt.Sprintf("%d", config.GinkgoConfig.ParallelNode)
-	nodeTmpDir, err := ioutil.TempDir("", "epinio-"+nodeSuffix)
+	nodeSuffix = fmt.Sprintf("%d", GinkgoParallelProcess())
+	nodeTmpDir, err := os.MkdirTemp("", "epinio-"+nodeSuffix)
 	Expect(err).NotTo(HaveOccurred())
 
-	out, err := testenv.CopyEpinioConfig(nodeTmpDir)
+	out, err := testenv.CopyEpinioSettings(nodeTmpDir)
 	Expect(err).ToNot(HaveOccurred(), out)
-	os.Setenv("EPINIO_CONFIG", nodeTmpDir+"/epinio.yaml")
+	os.Setenv("EPINIO_SETTINGS", nodeTmpDir+"/epinio.yaml")
 
-	config, err := epinioConfig.LoadFrom(nodeTmpDir + "/epinio.yaml")
+	config, err := settings.LoadFrom(nodeTmpDir + "/epinio.yaml")
 	Expect(err).NotTo(HaveOccurred())
 
-	env = testenv.New(nodeTmpDir, testenv.Root(), config.User, config.Password)
+	env = testenv.New(nodeTmpDir, testenv.Root(), config.User, config.Password, "", "")
 
 	out, err = proc.Run(testenv.Root(), false, "kubectl", "get", "ingress",
 		"--namespace", "epinio", "epinio",
@@ -60,7 +68,7 @@ var _ = BeforeSuite(func() {
 
 var _ = AfterSuite(func() {
 	if !testenv.SkipCleanup() {
-		fmt.Printf("Deleting tmpdir on node %d\n", config.GinkgoConfig.ParallelNode)
+		fmt.Printf("Deleting tmpdir on node %d\n", GinkgoParallelProcess())
 		testenv.DeleteTmpDir(nodeTmpDir)
 	}
 })
@@ -73,7 +81,7 @@ func FailWithReport(message string, callerSkip ...int) {
 	// NOTE: Use something like the following if you need to debug failed tests
 	// fmt.Println("\nA test failed. You may find the following information useful for debugging:")
 	// fmt.Println("The cluster pods: ")
-	// out, err := helpers.Kubectl("get pods --all-namespaces")
+	// out, err := proc.Kubectl("get pods --all-namespaces")
 	// if err != nil {
 	// 	fmt.Print(err.Error())
 	// } else {
