@@ -1,8 +1,18 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package testenv
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"strconv"
 	"time"
@@ -11,7 +21,7 @@ import (
 func AfterEachSleep() {
 	p := root + afterEachSleepPath
 	if _, err := os.Stat(p); err == nil {
-		if data, err := ioutil.ReadFile(p); err == nil {
+		if data, err := os.ReadFile(p); err == nil {
 			if s, err := strconv.Atoi(string(data)); err == nil {
 				t := time.Duration(s) * time.Second
 				fmt.Printf("Found '%s', sleeping for '%s'", p, t)
