@@ -1,7 +1,18 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package helpers
 
 import (
-	"io/ioutil"
+	"os"
 	"path"
 
 	"github.com/mholt/archiver/v3"
@@ -9,18 +20,17 @@ import (
 )
 
 func Tar(dir string) (string, string, error) {
-	files, err := ioutil.ReadDir(dir)
+	files, err := os.ReadDir(dir)
 	if err != nil {
 		return "", "", errors.Wrap(err, "cannot read the apps source files")
 	}
 	sources := []string{}
 	for _, f := range files {
-		// The FileInfo entries returned by ReadDir provide
-		// only the base name of the file or directory they
-		// are for. We have to add back the path of the
-		// application directory to get the proper paths to
-		// the files and directories to assemble in the
-		// tarball.
+		// The os.DirEntry structures returned by ReadDir() provide only the base
+		// name of the file or directory they are for. We have to add back the
+		// path of the application directory to get the proper paths to the files
+		// and directories to assemble in the tarball.
+
 		// Ignore git config files in the app sources.
 		if f.Name() == ".git" || f.Name() == ".gitignore" || f.Name() == ".gitmodules" || f.Name() == ".gitconfig" || f.Name() == ".git-credentials" {
 			continue
@@ -29,7 +39,7 @@ func Tar(dir string) (string, string, error) {
 	}
 
 	// create a tmpDir - tarball dir and POST
-	tmpDir, err := ioutil.TempDir("", "epinio-app")
+	tmpDir, err := os.MkdirTemp("", "epinio-app")
 	if err != nil {
 		return "", "", errors.Wrap(err, "can't create temp directory")
 	}
