@@ -1,38 +1,69 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package acceptance_test
 
 import (
 	"github.com/epinio/epinio/acceptance/helpers/catalog"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/epinio/epinio/acceptance/helpers/matchers"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Bounds between Apps & Services", func() {
-	var org string
-	dockerImageURL := "splatform/sample-app"
+var _ = Describe("Bounds between Apps & Configurations", LApplication, func() {
+	var namespace string
+	containerImageURL := "epinio/sample-app"
 
 	BeforeEach(func() {
-		org = catalog.NewOrgName()
-		env.SetupAndTargetOrg(org)
+		namespace = catalog.NewNamespaceName()
+		env.SetupAndTargetNamespace(namespace)
 	})
+
+	AfterEach(func() {
+		env.DeleteNamespace(namespace)
+	})
+
 	Describe("Display", func() {
 		var appName string
-		var serviceName string
+		var configurationName string
+
 		BeforeEach(func() {
 			appName = catalog.NewAppName()
-			serviceName = catalog.NewServiceName()
+			configurationName = catalog.NewConfigurationName()
 
-			env.MakeDockerImageApp(appName, 1, dockerImageURL)
-			env.MakeService(serviceName)
-			env.BindAppService(appName, serviceName, org)
+			env.MakeContainerImageApp(appName, 1, containerImageURL)
+			env.MakeConfiguration(configurationName)
+			env.BindAppConfiguration(appName, configurationName, namespace)
 		})
-		It("shows the bound app for services list, and vice versa", func() {
-			out, err := env.Epinio("", "service", "list")
+
+		AfterEach(func() {
+			// Delete app first, as this also unbinds the configuration
+			env.CleanupApp(appName)
+			env.CleanupConfiguration(configurationName)
+		})
+
+		It("shows the bound app for configurations list, and vice versa", func() {
+			out, err := env.Epinio("", "configuration", "list")
 			Expect(err).ToNot(HaveOccurred(), out)
-			Expect(out).To(MatchRegexp(serviceName + `.*` + appName))
+
+			Expect(out).To(
+				HaveATable(
+					WithHeaders("NAME", "CREATED", "TYPE", "ORIGIN", "APPLICATIONS"),
+					WithRow(configurationName, WithDate(), "custom", "", appName),
+				),
+			)
 
 			// The next check uses `Eventually` because binding the
-			// service to the app forces a restart of the app's
+			// configuration to the app forces a restart of the app's
 			// pod. It takes the system some time to terminate the
 			// old pod, and spin up the new, during which `app list`
 			// will return inconsistent results about the desired
@@ -43,12 +74,12 @@ var _ = Describe("Bounds between Apps & Services", func() {
 				out, err := env.Epinio("", "app", "list")
 				Expect(err).ToNot(HaveOccurred(), out)
 				return out
-			}, "5m").Should(MatchRegexp(appName + `.*\|.*1\/1.*\|.*` + serviceName))
-		})
-		AfterEach(func() {
-			// Delete app first, as this also unbinds the service
-			env.CleanupApp(appName)
-			env.CleanupService(serviceName)
+			}, "5m").Should(
+				HaveATable(
+					WithHeaders("NAME", "CREATED", "STATUS", "ROUTES", "CONFIGURATIONS", "STATUS DETAILS"),
+					WithRow(appName, WithDate(), "1/1", appName+".*", configurationName, ""),
+				),
+			)
 		})
 	})
 })
