@@ -1,28 +1,36 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package proc
 
 import (
 	"bytes"
 	"io"
 	"os"
+	"os/exec"
 
-	"github.com/codeskyblue/kexec"
+	"github.com/pkg/errors"
 )
 
-func Get(dir, command string, arg ...string) (*kexec.KCommand, error) {
-	var commandDir string
+func Get(dir, command string, arg ...string) (*exec.Cmd, error) {
 	var err error
 
 	if dir == "" {
-		commandDir, err = os.Getwd()
-		if err != nil {
+		if dir, err = os.Getwd(); err != nil {
 			return nil, err
 		}
-	} else {
-		commandDir = dir
 	}
 
-	p := kexec.Command(command, arg...)
-	p.Dir = commandDir
+	p := exec.Command(command, arg...)
+	p.Dir = dir
 
 	return p, nil
 }
@@ -32,25 +40,36 @@ func RunW(cmd string, args ...string) (string, error) {
 	return Run("", false, cmd, args...)
 }
 
-func Run(dir string, toStdout bool, cmd string, arg ...string) (string, error) {
-	p, err := Get(dir, cmd, arg...)
+func Run(dir string, toStdout bool, command string, args ...string) (string, error) {
+	cmd := exec.Command(command, args...)
+
+	var b bytes.Buffer
+	if toStdout {
+		cmd.Stdout = io.MultiWriter(os.Stdout, &b)
+		cmd.Stderr = io.MultiWriter(os.Stderr, &b)
+	} else {
+		cmd.Stdout = &b
+		cmd.Stderr = &b
+	}
+
+	cmd.Dir = dir
+
+	err := cmd.Run()
+	return b.String(), err
+}
+
+// Kubectl invokes the `kubectl` command in PATH, running the specified command.
+// It returns the command output and/or error.
+func Kubectl(command ...string) (string, error) {
+	_, err := exec.LookPath("kubectl")
+	if err != nil {
+		return "", errors.Wrap(err, "kubectl not in path")
+	}
+
+	currentdir, err := os.Getwd()
 	if err != nil {
 		return "", err
 	}
 
-	var b bytes.Buffer
-	if toStdout {
-		p.Stdout = io.MultiWriter(os.Stdout, &b)
-		p.Stderr = io.MultiWriter(os.Stderr, &b)
-	} else {
-		p.Stdout = &b
-		p.Stderr = &b
-	}
-
-	if err := p.Run(); err != nil {
-		return b.String(), err
-	}
-
-	err = p.Wait()
-	return b.String(), err
+	return Run(currentdir, false, "kubectl", command...)
 }
