@@ -1,44 +1,78 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package install_test
 
 import (
 	"encoding/json"
 	"os"
+	"time"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/epinio/epinio/acceptance/helpers/catalog"
 	"github.com/epinio/epinio/acceptance/helpers/epinio"
 	"github.com/epinio/epinio/acceptance/helpers/proc"
 	"github.com/epinio/epinio/acceptance/helpers/route53"
 	"github.com/epinio/epinio/acceptance/testenv"
 )
 
-var _ = Describe("<Scenario5> Azure, Letsencrypt", func() {
+var _ = Describe("<Scenario5> Azure, Letsencrypt-staging, External Registry", func() {
 	var (
-		appName      = catalog.NewAppName()
-		domain       string
-		epinioHelper epinio.Epinio
-		flags        []string
-		loadbalancer string
-		zoneID       string
+		appName          string
+		domain           string
+		epinioHelper     epinio.Epinio
+		flags            []string
+		loadbalancer     string
+		registryUsername string
+		registryPassword string
+		zoneID           string
+		extraEnvName     string
+		extraEnvValue    string
+		name_exists      bool
+		value_exists     bool
 	)
 
 	BeforeEach(func() {
 		epinioHelper = epinio.NewEpinioHelper(testenv.EpinioBinaryPath())
 
 		// use Route53
-		domain = os.Getenv("AKS_DOMAIN")
+		domain = os.Getenv("EPINIO_SYSTEM_DOMAIN")
 		Expect(domain).ToNot(BeEmpty())
 
 		zoneID = os.Getenv("AWS_ZONE_ID")
 		Expect(zoneID).ToNot(BeEmpty())
 
+		appName = "external-reg-test-aks-letsencrypt"
+
+		registryUsername = os.Getenv("REGISTRY_USERNAME")
+		Expect(registryUsername).ToNot(BeEmpty())
+
+		registryPassword = os.Getenv("REGISTRY_PASSWORD")
+		Expect(registryPassword).ToNot(BeEmpty())
 		flags = []string{
-			"--skip-default-namespace",
-			"--tls-issuer=letsencrypt-production",
-			"--use-internal-registry-node-port=false",
-			"--system-domain=" + domain,
+			"--set", "server.disableTracking=true", // disable tracking during tests
+			"--set", "global.domain=" + domain,
+			"--set", "global.tlsIssuer=letsencrypt-staging", // let epinio create the ClusterIssuer
+			"--set", "containerregistry.enabled=false",
+			"--set", "global.registryURL=registry.hub.docker.com",
+			"--set", "global.registryUsername=" + registryUsername,
+			"--set", "global.registryPassword=" + registryPassword,
+			"--set", "global.registryNamespace=epinio",
+		}
+
+		extraEnvName, name_exists = os.LookupEnv("EXTRAENV_NAME")
+		extraEnvValue, value_exists = os.LookupEnv("EXTRAENV_VALUE")
+		if name_exists && value_exists {
+			flags = append(flags, "--set", "extraEnv[0].name="+extraEnvName, "--set-string", "extraEnv[0].value="+extraEnvValue)
 		}
 	})
 
@@ -47,64 +81,105 @@ var _ = Describe("<Scenario5> Azure, Letsencrypt", func() {
 		Expect(err).NotTo(HaveOccurred(), out)
 	})
 
-	It("installs and passes scenario", func() {
-		By("Installing Traefik", func() {
-			out, err := epinioHelper.Run("install-ingress")
-			Expect(err).NotTo(HaveOccurred(), out)
-			Expect(out).To(Or(ContainSubstring("Traefik deployed"), ContainSubstring("Traefik Ingress info")))
-		})
+	It("Installs with letsencrypt-staging cert and pushes an app", func() {
+		// Do not create the letsencrypt-staging clusterIssuer here, epinio will create it instead
+		By("Checking LoadBalancer IP", func() {
+			// Ensure that Traefik LB is not in Pending state anymore, could take time
+			Eventually(func() string {
+				out, err := proc.RunW("kubectl", "get", "svc", "-n", "traefik", "traefik", "--no-headers")
+				Expect(err).NotTo(HaveOccurred(), out)
+				return out
+			}, "4m", "2s").ShouldNot(ContainSubstring("<pending>"))
 
-		By("Extracting AKS Loadbalancer Name", func() {
 			out, err := proc.RunW("kubectl", "get", "service", "-n", "traefik", "traefik", "-o", "json")
 			Expect(err).NotTo(HaveOccurred(), out)
 
+			// Check that an IP address for LB is configured
 			status := &testenv.LoadBalancerHostname{}
 			err = json.Unmarshal([]byte(out), status)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(status.Status.LoadBalancer.Ingress).To(HaveLen(1))
 			loadbalancer = status.Status.LoadBalancer.Ingress[0].IP
-			Expect(loadbalancer).ToNot(BeEmpty(), out)
+			Expect(loadbalancer).ToNot(BeEmpty())
 		})
 
 		By("Updating DNS Entries", func() {
-			change := route53.A(domain, loadbalancer)
-			out, err := route53.Upsert(zoneID, change, nodeTmpDir)
+			change := route53.A(domain, loadbalancer, "UPSERT")
+			out, err := route53.Update(zoneID, change, nodeTmpDir)
 			Expect(err).NotTo(HaveOccurred(), out)
 
-			change = route53.A("*."+domain, loadbalancer)
-			out, err = route53.Upsert(zoneID, change, nodeTmpDir)
+			change = route53.A("*."+domain, loadbalancer, "UPSERT")
+			out, err = route53.Update(zoneID, change, nodeTmpDir)
 			Expect(err).NotTo(HaveOccurred(), out)
 		})
+
+		// Check that DNS entry is correctly propagated
+		By("Checking that DNS entry is correctly propagated", func() {
+			Eventually(func() string {
+				out, err := route53.TestDnsAnswer(zoneID, domain, "A")
+				Expect(err).NotTo(HaveOccurred(), out)
+
+				answer := &route53.DNSAnswer{}
+				err = json.Unmarshal([]byte(out), answer)
+				Expect(err).NotTo(HaveOccurred())
+				if len(answer.RecordData) == 0 {
+					return ""
+				}
+				return answer.RecordData[0]
+			}, "5m", "2s").Should(Equal(loadbalancer))
+		})
+
+		// Workaround to (try to!) ensure that the DNS is really propagated!
+		time.Sleep(3 * time.Minute)
 
 		By("Installing Epinio", func() {
 			out, err := epinioHelper.Install(flags...)
 			Expect(err).NotTo(HaveOccurred(), out)
-			Expect(out).To(ContainSubstring("Epinio installed."))
+			Expect(out).To(ContainSubstring("STATUS: deployed"))
 
 			out, err = testenv.PatchEpinio()
 			Expect(err).ToNot(HaveOccurred(), out)
 		})
 
-		// Now create the default org which we skipped because
-		// it would fail before patching.
-		testenv.EnsureDefaultWorkspace(testenv.EpinioBinaryPath())
-		out, err := epinioHelper.Run("target", testenv.DefaultWorkspace)
-		Expect(err).ToNot(HaveOccurred(), out)
+		By("Connecting to Epinio", func() {
+			Eventually(func() string {
+				out, _ := epinioHelper.Run("login", "-u", "admin", "-p", "password", "--trust-ca", "https://epinio."+domain)
+				return out
+			}, "2m", "5s").Should(ContainSubstring("Login successful"))
+		})
 
-		Eventually(func() string {
-			out, _ := epinioHelper.Run("info")
-			return out
-		}).Should(ContainSubstring("Epinio Version: "))
+		By("Checking Epinio info command", func() {
+			Eventually(func() string {
+				out, _ := epinioHelper.Run("info")
+				return out
+			}, "2m", "2s").Should(ContainSubstring("Epinio Server Version:"))
+		})
 
 		By("Pushing an app", func() {
-			out, err = epinioHelper.Run("push", appName, testenv.AssetPath("sample-app"))
+			out, err := epinioHelper.Run("push",
+				"--name", appName,
+				"--path", testenv.AssetPath("sample-app"))
 			Expect(err).NotTo(HaveOccurred(), out)
 
-			// verify cluster_issuer is used
+			// Verify cluster_issuer is used
 			out, err = proc.RunW("kubectl", "get", "certificate",
-				"-n", testenv.DefaultWorkspace, appName, "-o", "jsonpath='{.spec.issuerRef.name}'")
+				"-n", testenv.DefaultWorkspace,
+				"--selector", "app.kubernetes.io/name="+appName,
+				"-o", "jsonpath='{.items[*].spec.issuerRef.name}'")
 			Expect(err).NotTo(HaveOccurred(), out)
-			Expect(out).To(Equal("'letsencrypt-production'"))
+			Expect(out).To(Equal("'letsencrypt-staging'"))
+
+			// Wait until all the CertRequests in all namespaces are Ready -> the CSRs have been approved and signed
+			out, err = proc.RunW("kubectl", "wait", "--for=condition=Ready",
+				"certificaterequest", "--selector", "app.kubernetes.io/managed-by=Helm",
+				"--all-namespaces", "--timeout=120s")
+			Expect(err).NotTo(HaveOccurred(), out)
+		})
+
+		By("Delete an app", func() {
+			out, err := epinioHelper.Run("apps", "delete", appName)
+			Expect(err).NotTo(HaveOccurred(), out)
+			Expect(out).To(Or(ContainSubstring("Applications Removed")))
 		})
 	})
 })
