@@ -1,11 +1,22 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package testenv
 
 import (
 	"fmt"
 	"os"
 
-	"github.com/epinio/epinio/helpers"
-	"github.com/onsi/ginkgo/config"
+	"github.com/epinio/epinio/acceptance/helpers/proc"
+	. "github.com/onsi/ginkgo/v2"
 )
 
 // RegistryUsername returns the docker registry username from the env
@@ -21,8 +32,8 @@ func RegistryPassword() string {
 // CreateRegistrySecret creates the docker registry image pull secret
 func CreateRegistrySecret() {
 	if RegistryUsername() != "" && RegistryPassword() != "" {
-		fmt.Printf("Creating image pull secret for Dockerhub on node %d\n", config.GinkgoConfig.ParallelNode)
-		_, _ = helpers.Kubectl("create", "secret", "docker-registry", "regcred",
+		fmt.Printf("Creating image pull secret for Dockerhub on node %d\n", GinkgoParallelProcess())
+		_, _ = proc.Kubectl("create", "secret", "docker-registry", "regcred",
 			"--docker-server", "https://index.docker.io/v1/",
 			"--docker-username", RegistryUsername(),
 			"--docker-password", RegistryPassword(),
