@@ -1,3 +1,14 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package testenv
 
 import (
@@ -48,7 +59,7 @@ func CleanupTmp() (string, error) {
 	return proc.Run("", true, "rm", append([]string{"-rf"}, temps...)...)
 }
 
-// CopyEpinioConfig copies the epinio yaml to the given dir
-func CopyEpinioConfig(dir string) (string, error) {
+// CopyEpinioSettings copies the epinio yaml to the given dir
+func CopyEpinioSettings(dir string) (string, error) {
 	return proc.Run("", false, "cp", EpinioYAML(), dir+"/epinio.yaml")
 }
