@@ -1,3 +1,14 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package config
 
 import (
@@ -13,7 +24,8 @@ import (
 // KubeConfigFlags adds a kubeconfig flag to the set
 func KubeConfigFlags(pf *pflag.FlagSet, argToEnv map[string]string) {
 	pf.StringP("kubeconfig", "c", "", "path to a kubeconfig, not required in-cluster")
-	viper.BindPFlag("kubeconfig", pf.Lookup("kubeconfig"))
+	err := viper.BindPFlag("kubeconfig", pf.Lookup("kubeconfig"))
+	checkErr(err)
 	argToEnv["kubeconfig"] = "KUBECONFIG"
 }
 
@@ -26,5 +38,9 @@ func KubeConfig() (*rest.Config, error) {
 	if err := NewChecker().Check(restConfig); err != nil {
 		return nil, errors.Wrap(err, "couldn't check kubeconfig; ensure kubeconfig is correct to continue")
 	}
+
+	restConfig.QPS = float32(viper.GetFloat64("kube-api-qps"))
+	restConfig.Burst = viper.GetInt("kube-api-burst")
+
 	return restConfig, nil
 }
