@@ -1,3 +1,14 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 // tailer manages objects which tail the logs of a collection of pods specified by a label selector.
 // This is similar to what the cli tool `stern` does.
 package tailer
@@ -10,14 +21,13 @@ import (
 	"time"
 
 	"github.com/epinio/epinio/helpers/kubernetes"
-	"github.com/epinio/epinio/helpers/tracelog"
+	"github.com/epinio/epinio/internal/cli/server/requestctx"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 )
 
-// Config contains the config for stern
 type Config struct {
 	Namespace             string           // Name of the namespace to monitor
 	PodQuery              *regexp.Regexp   // Limit monitoring to pods matching the RE
@@ -32,6 +42,7 @@ type Config struct {
 	LabelSelector         labels.Selector
 	TailLines             *int64
 	Template              *template.Template // Template to apply to log entries for formatting
+	Ordered               bool               // Featch/stream logs in container order, synchronously
 }
 
 // ContainerLogLine is an object that represents a line from the logs of a container.
@@ -46,7 +57,7 @@ type ContainerLogLine struct {
 // FetchLogs writes all the logs of the matching containers to the logChan.
 // If ctx is Done() the method stops even if not all logs are fetched.
 func FetchLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.WaitGroup, config *Config, cluster *kubernetes.Cluster) error {
-	logger := tracelog.NewLogger().WithName("fetching-logs").V(3)
+	logger := requestctx.Logger(ctx).WithName("fetching-logs").V(3)
 	var namespace string
 	if config.AllNamespaces {
 		namespace = ""
@@ -54,6 +65,7 @@ func FetchLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wait
 		return errors.New("no namespace set for tailing logs")
 	}
 
+	logger.Info("list pods")
 	podList, err := cluster.Kubectl.CoreV1().Pods(namespace).List(
 		ctx, metav1.ListOptions{LabelSelector: config.LabelSelector.String()})
 	if err != nil {
@@ -63,7 +75,7 @@ func FetchLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wait
 	tails := []*Tail{}
 	newTail := func(pod corev1.Pod, c corev1.Container) *Tail {
 		return NewTail(pod.Namespace, pod.Name, c.Name,
-			tracelog.NewLogger().WithName("log-tracing").V(4),
+			requestctx.Logger(ctx).WithName("log-tracing").V(4),
 			cluster.Kubectl,
 			&TailOptions{
 				Timestamps:   config.Timestamps,
@@ -86,22 +98,45 @@ func FetchLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wait
 		return true
 	}
 
+	logger.Info("filter pods, containers")
+
 	for _, pod := range podList.Items {
-		for _, c := range pod.Spec.Containers {
-			if !acceptable(c) {
-				continue
-			}
-			tails = append(tails, newTail(pod, c))
-		}
 		for _, c := range pod.Spec.InitContainers {
 			if !acceptable(c) {
 				continue
 			}
 			tails = append(tails, newTail(pod, c))
+
+			logger.Info("have", "namespace", pod.Namespace, "pod", pod.Name, "container", c.Name)
+		}
+		for _, c := range pod.Spec.Containers {
+			if !acceptable(c) {
+				continue
+			}
+			tails = append(tails, newTail(pod, c))
+
+			logger.Info("have", "namespace", pod.Namespace, "pod", pod.Name, "container", c.Name)
 		}
 	}
 
+	if config.Ordered {
+		logger.Info("fetch in order")
+
+		for _, t := range tails {
+			logger.Info("tail", "namespace", t.Namespace, "pod", t.PodName, "container", t.ContainerName)
+
+			err := t.Start(ctx, logChan, false)
+			if err != nil {
+				logger.Error(err, "failed to start a Tail")
+			}
+		}
+
+		return nil
+	}
+
 	for _, t := range tails {
+		logger.Info("tail", "namespace", t.Namespace, "pod", t.PodName, "container", t.ContainerName)
+
 		wg.Add(1)
 		go func(tail *Tail) {
 			err := tail.Start(ctx, logChan, false)
@@ -119,7 +154,7 @@ func FetchLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wait
 // logChan.  The containers are determined by an internal watcher
 // polling the cluster for pod __changes__.
 func StreamLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.WaitGroup, config *Config, cluster *kubernetes.Cluster) error {
-	logger := tracelog.NewLogger().WithName("tail-handling").V(3)
+	logger := requestctx.Logger(ctx).WithName("tail-handling").V(3)
 
 	var namespace string
 	if config.AllNamespaces {
@@ -132,7 +167,6 @@ func StreamLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wai
 		"pods", config.PodQuery.String(),
 		"containers", config.ContainerQuery.String(),
 		"excluded", config.ExcludeContainerQuery.String(),
-		"state", config.ContainerState,
 		"selector", config.LabelSelector.String())
 	added, removed, err := Watch(ctx, cluster.Kubectl.CoreV1().Pods(namespace),
 		config.PodQuery, config.ContainerQuery, config.ExcludeContainerQuery, config.ContainerState, config.LabelSelector)
@@ -153,13 +187,13 @@ func StreamLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wai
 		case p := <-added:
 			id := p.GetID()
 			if tails[id] != nil {
-				break
+				continue
 			}
 
 			logger.Info("tailer add", "id", id)
 
 			tail := NewTail(p.Namespace, p.Pod, p.Container,
-				tracelog.NewLogger().WithName("log-tracing"),
+				requestctx.Logger(ctx).WithName("log-tracing"),
 				cluster.Kubectl,
 				&TailOptions{
 					Timestamps:   config.Timestamps,
@@ -181,15 +215,20 @@ func StreamLogs(ctx context.Context, logChan chan ContainerLogLine, wg *sync.Wai
 				logger.Info("tailer done", "id", id)
 				wg.Done()
 			}(id)
+
+			logger.Info("tailer added", "id", id)
+
 		case p := <-removed:
 			id := p.GetID()
 			if tails[id] == nil {
-				break
+				continue
 			}
 
 			logger.Info("tailer remove", "id", id)
 
 			delete(tails, id)
+
+			logger.Info("tailer removed", "id", id)
 		case <-ctx.Done():
 			logger.Info("received stop request")
 			return nil
