@@ -1,14 +1,23 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package application
 
 import (
 	"context"
-	"math"
 	"strconv"
 
 	"github.com/epinio/epinio/helpers/kubernetes"
 	"github.com/epinio/epinio/pkg/api/core/v1/models"
 	v1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/util/retry"
 )
@@ -24,17 +33,24 @@ func Scaling(ctx context.Context, cluster *kubernetes.Cluster, appRef models.App
 		return 0, err
 	}
 
-	result, err := strconv.Atoi(string(scaleSecret.Data[instanceKey])) // nolint:gosec // overflow blocked by guards
+	return ScalingFromSecret(scaleSecret)
+}
+
+// ScalingFromSecret is the core of Scaling, extracting the desired number of instances from the
+// secret containing them.
+func ScalingFromSecret(scaleSecret *v1.Secret) (int32, error) {
+	i, err := strconv.ParseInt(string(scaleSecret.Data[instanceKey]), 10, 32)
 	if err != nil {
 		return 0, err
 	}
+	result := int32(i)
 
 	// Reject bad values, and assume single instance - Return err better ? Save back, fix resource ?
-	if result <= 0 || result > math.MaxInt32 {
+	if result < 0 {
 		result = 1
 	}
 
-	return int32(result), nil
+	return result, nil
 }
 
 // ScalingSet sets the desired number of instances for the named application.
@@ -65,7 +81,7 @@ func scaleUpdate(ctx context.Context, cluster *kubernetes.Cluster,
 
 		modifyScaling(scaleSecret)
 
-		_, err = cluster.Kubectl.CoreV1().Secrets(appRef.Org).Update(
+		_, err = cluster.Kubectl.CoreV1().Secrets(appRef.Namespace).Update(
 			ctx, scaleSecret, metav1.UpdateOptions{})
 
 		return err
@@ -76,50 +92,5 @@ func scaleUpdate(ctx context.Context, cluster *kubernetes.Cluster,
 // instances. If necessary it creates that secret.
 func scaleLoad(ctx context.Context, cluster *kubernetes.Cluster, appRef models.AppRef) (*v1.Secret, error) {
 	secretName := appRef.MakeScaleSecretName()
-
-	scaleSecret, err := cluster.GetSecret(ctx, appRef.Org, secretName)
-	if err != nil {
-		if !apierrors.IsNotFound(err) {
-			return nil, err
-		}
-
-		// Error is `Not Found`. Create the secret.
-
-		app, err := Get(ctx, cluster, appRef)
-		if err != nil {
-			// Should not happen. The application was validated to exist already somewhere
-			// by this function's callers.
-			return nil, err
-		}
-
-		owner := metav1.OwnerReference{
-			APIVersion: app.GetAPIVersion(),
-			Kind:       app.GetKind(),
-			Name:       app.GetName(),
-			UID:        app.GetUID(),
-		}
-
-		scaleSecret = &v1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      secretName,
-				Namespace: appRef.Org,
-				OwnerReferences: []metav1.OwnerReference{
-					owner,
-				},
-				Labels: map[string]string{
-					"app.kubernetes.io/name":       appRef.Name,
-					"app.kubernetes.io/part-of":    appRef.Org,
-					"app.kubernetes.io/managed-by": "epinio",
-					"app.kubernetes.io/component":  "application",
-				},
-			},
-		}
-		err = cluster.CreateSecret(ctx, appRef.Org, *scaleSecret)
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return scaleSecret, nil
+	return loadOrCreateSecret(ctx, cluster, appRef, secretName, "scaling")
 }
