@@ -1,3 +1,14 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package usercmd
 
 import (
@@ -5,25 +16,23 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"sort"
+	"strconv"
 	"strings"
-	"sync"
 
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/util/validation"
 
 	"github.com/epinio/epinio/helpers"
+	"github.com/epinio/epinio/internal/cli/termui"
 	"github.com/epinio/epinio/internal/duration"
+	"github.com/epinio/epinio/pkg/api/core/v1/client"
 	"github.com/epinio/epinio/pkg/api/core/v1/models"
 )
 
 type PushParams struct {
-	Configuration models.ApplicationUpdateRequest // instances, services, EVs
-	Docker        string
-	GitRev        string
-	BuilderImage  string
-	Name          string
-	Path          string
+	models.ApplicationManifest
 }
 
 // Push pushes an app
@@ -31,43 +40,65 @@ type PushParams struct {
 // * upload
 // * stage
 // * (tail logs)
-// * wait for pipelinerun
+// * wait for staging to be done (complete or fail)
 // * deploy
 // * wait for app
-func (c *EpinioClient) Push(ctx context.Context, params PushParams) error {
-	name := params.Name
-	source := params.Path
+func (c *EpinioClient) AppPush(ctx context.Context, manifest models.ApplicationManifest) error { // nolint: gocyclo // Many ifs for view purposes
 
-	appRef := models.AppRef{Name: name, Org: c.Config.Org}
+	// Use settings default if user did not specify --app-chart
+	if manifest.Configuration.AppChart == "" {
+		manifest.Configuration.AppChart = c.Settings.AppChart
+	}
+
+	source := manifest.Origin.String()
+	appRef := models.AppRef{
+		Meta: models.Meta{
+			Name:      manifest.Name,
+			Namespace: c.Settings.Namespace,
+		},
+	}
 	log := c.Log.
 		WithName("Push").
 		WithValues("Name", appRef.Name,
-			"Namespace", appRef.Org,
+			"Namespace", appRef.Namespace,
 			"Sources", source)
 	log.Info("start")
 	defer log.Info("return")
 	details := log.V(1) // NOTE: Increment of level, not absolute. Visible via TRACE_LEVEL=2
 
-	sourceToShow := source
-	if params.GitRev != "" {
-		sourceToShow = fmt.Sprintf("%s @ %s", sourceToShow, params.GitRev)
-	}
-
 	msg := c.ui.Note().
+		WithStringValue("Manifest", manifest.Self). // This path is already platform-specific
 		WithStringValue("Name", appRef.Name).
-		WithStringValue("Sources", sourceToShow).
-		WithStringValue("Namespace", appRef.Org)
+		WithStringValue("Source Origin", source).
+		WithStringValue("AppChart", manifest.Configuration.AppChart).
+		WithStringValue("Target Namespace", appRef.Namespace)
+	for _, ev := range manifest.Configuration.Environment.List() {
+		msg = msg.WithStringValue(fmt.Sprintf("Environment '%s'", ev.Name), ev.Value)
+	}
+	// TODO ? Make this a table for nicer alignment
 
 	if err := c.TargetOk(); err != nil {
 		return err
 	}
 
-	if len(params.Configuration.Services) > 0 {
-		msg = msg.WithStringValue("Services:",
-			strings.Join(params.Configuration.Services, ", "))
+	// Show builder, if relevant (i.e. path/git sources, not for container)
+	if manifest.Origin.Kind != models.OriginContainer &&
+		manifest.Staging.Builder != "" {
+		msg = msg.WithStringValue("Builder", manifest.Staging.Builder)
 	}
 
-	msg.Msg("About to push an application with given name and sources into the specified namespace")
+	if manifest.Configuration.Instances != nil {
+		msg = msg.WithStringValue("Instances",
+			strconv.Itoa(int(*manifest.Configuration.Instances)))
+	}
+	if len(manifest.Configuration.Configurations) > 0 {
+		msg = msg.WithStringValue("Configurations",
+			strings.Join(manifest.Configuration.Configurations, ", "))
+	}
+
+	msg = routeMessage(msg, manifest.Configuration.Routes)
+
+	msg.Msg("About to push an application with the given setup")
 
 	c.ui.Exclamation().
 		Timeout(duration.UserAbort()).
@@ -82,83 +113,99 @@ func (c *EpinioClient) Push(ctx context.Context, params PushParams) error {
 	// AppCreate
 	c.ui.Normal().Msg("Create the application resource ...")
 
-	request := models.ApplicationCreateRequest{
+	updateRequest := models.NewApplicationUpdateRequest(manifest)
+	_, err := c.API.AppCreate(models.ApplicationCreateRequest{
 		Name:          appRef.Name,
-		Configuration: params.Configuration,
-	}
-
-	_, err := c.API.AppCreate(request, appRef.Org)
+		Configuration: updateRequest,
+	}, appRef.Namespace)
 	if err != nil {
 		// try to recover if it's a response type Conflict error and not a http connection error
-		rerr, ok := err.(interface{ StatusCode() int })
-
-		if !ok {
+		epinioAPIError := &client.APIError{}
+		if !errors.As(err, &epinioAPIError) {
 			return err
 		}
 
-		if rerr.StatusCode() != http.StatusConflict {
+		if epinioAPIError.StatusCode != http.StatusConflict {
 			return err
 		}
 
 		c.ui.Normal().Msg("Application exists, updating ...")
 		details.Info("app exists conflict")
 
-		_, err := c.API.AppUpdate(params.Configuration, appRef.Org, appRef.Name)
+		// do not restart during a push
+		restart := false
+		updateRequest.Restart = &restart
+
+		_, err := c.API.AppUpdate(updateRequest, appRef.Namespace, appRef.Name)
 		if err != nil {
 			return err
 		}
 	}
 
+	// check customization
+	_, err = c.API.AppValidateCV(appRef.Namespace, appRef.Name)
+	if err != nil {
+		return err
+	}
+
 	// AppUpload / AppImportGit
 	var blobUID string
-	if params.GitRev == "" && params.Docker == "" {
-		c.ui.Normal().Msg("Collecting the application sources ...")
-
-		tmpDir, tarball, err := helpers.Tar(source)
-		defer func() {
-			if tmpDir != "" {
-				_ = os.RemoveAll(tmpDir)
-			}
-		}()
+	switch manifest.Origin.Kind {
+	case models.OriginNone:
+		return fmt.Errorf("%s", "No application origin")
+	case models.OriginPath:
+		uploadedSourceBlobID, err := c.uploadSources(log, appRef, source)
 		if err != nil {
 			return err
 		}
-
-		c.ui.Normal().Msg("Uploading application code ...")
-
-		details.Info("upload code")
-		upload, err := c.API.AppUpload(appRef.Org, appRef.Name, tarball)
-		if err != nil {
-			return err
-		}
-		log.V(3).Info("upload response", "response", upload)
-
-		blobUID = upload.BlobUID
-
-	} else if params.GitRev != "" {
+		blobUID = uploadedSourceBlobID
+	case models.OriginGit:
 		c.ui.Normal().Msg("Importing the application sources from Git ...")
 
-		gitRef := models.GitRef{
-			URL:      source,
-			Revision: params.GitRev,
+		gitOrigin := manifest.Origin.Git
+		if gitOrigin == nil {
+			return errors.New("git origin is nil")
 		}
-		response, err := c.API.AppImportGit(appRef, gitRef)
+
+		// validate provider reference, if actually present (git origin, and specified)
+		if gitOrigin.Provider != "" {
+			if _, err := models.GitProviderFromString(string(gitOrigin.Provider)); err != nil {
+				return errors.Wrapf(err, "bad git provider `%s`", gitOrigin.Provider)
+			}
+
+			if err := gitOrigin.Provider.ValidateURL(gitOrigin.URL); err != nil {
+				return errors.Wrap(err, "validating git url")
+			}
+		}
+
+		response, err := c.API.AppImportGit(appRef.Namespace, appRef.Name, *gitOrigin)
 		if err != nil {
 			return errors.Wrap(err, "importing git remote")
 		}
+
 		blobUID = response.BlobUID
+		// if the server resolved the branch use that in the Git Origin
+		if response.Branch != "" {
+			manifest.Origin.Git.Branch = response.Branch
+		}
+		// update the revision with the one updated by the server
+		manifest.Origin.Git.Revision = response.Revision
+
+	case models.OriginContainer:
+		// Nothing to upload (nor stage)
 	}
 
 	// AppStage
 	stageID := ""
 	var stageResponse *models.StageResponse
-	if params.Docker == "" {
-		c.ui.Normal().Msg("Staging application via docker image ...")
+	if manifest.Origin.Kind != models.OriginContainer {
+		c.ui.Normal().Msg("Staging application with code...")
+		c.ui.ProgressNote().Msg("Running staging")
 
 		req := models.StageRequest{
 			App:          appRef,
 			BlobUID:      blobUID,
-			BuilderImage: params.BuilderImage,
+			BuilderImage: manifest.Staging.Builder,
 		}
 		details.Info("staging code", "Blob", blobUID)
 		stageResponse, err = c.API.AppStage(req)
@@ -169,25 +216,37 @@ func (c *EpinioClient) Push(ctx context.Context, params PushParams) error {
 		log.V(3).Info("stage response", "response", stageResponse)
 
 		details.Info("start tailing logs", "StageID", stageResponse.Stage.ID)
-		err = c.stageLogs(details, appRef, stageResponse.Stage.ID)
+		c.stageLogs(appRef, stageResponse.Stage.ID)
+
+		details.Info("wait for job", "StageID", stageID)
+
+		// blocking function that wait until the staging is done
+		err = stagingWithRetry(log.V(1), c.API, appRef.Namespace, stageID)
 		if err != nil {
-			return err
+			c.ui.Note().Msgf(
+				"You can access the staging logs at any time, either in the UI or with the CLI using this command:\n\nepinio app logs --staging %s",
+				appRef.Name)
+			return errors.Wrap(err, "waiting for staging failed")
 		}
 	}
 
 	// AppDeploy
 	c.ui.Normal().Msg("Deploying application ...")
 	deployRequest := models.DeployRequest{
-		App: appRef,
+		App:    appRef,
+		Origin: manifest.Origin,
 	}
-	// If docker param is specified, then we just take it into ImageURL
+	// If container param is specified, then we just take it into ImageURL
 	// If not, we take the one from the staging response
-	if params.Docker != "" {
-		deployRequest.ImageURL = params.Docker
+	if manifest.Origin.Kind == models.OriginContainer {
+		deployRequest.ImageURL = manifest.Origin.Container
 	} else {
 		deployRequest.ImageURL = stageResponse.ImageURL
 		deployRequest.Stage = models.StageRef{ID: stageID}
 	}
+
+	s := c.ui.Progress("Waiting for deployment")
+	defer s.Stop()
 
 	deployResponse, err := c.API.AppDeploy(deployRequest)
 	if err != nil {
@@ -197,46 +256,93 @@ func (c *EpinioClient) Push(ctx context.Context, params PushParams) error {
 	details.Info("wait for application resources")
 	c.ui.ProgressNote().KeeplineUnder(1).Msg("Creating application resources")
 
-	_, err = c.API.AppRunning(appRef)
-	if err != nil {
-		return errors.Wrap(err, "waiting for app failed")
+	routes := []string{}
+	for _, d := range deployResponse.Routes {
+		routes = append(routes, fmt.Sprintf("https://%s", d))
 	}
 
-	c.ui.Success().
-		WithStringValue("Name", appRef.Name).
-		WithStringValue("Namespace", appRef.Org).
-		WithStringValue("Route", fmt.Sprintf("https://%s", deployResponse.Route)).
-		WithStringValue("Builder Image", params.BuilderImage).
-		Msg("App is online.")
-
+	c.reportOK(appRef, manifest.Staging.Builder, routes)
 	return nil
 }
 
-func (c *EpinioClient) stageLogs(details logr.Logger, appRef models.AppRef, stageID string) error {
-	// Buffered because the go routine may no longer be listening when we try
-	// to stop it. Stopping it should be a fire and forget. We have wg to wait
-	// for the routine to be gone.
-	stopChan := make(chan bool, 1)
-	var wg sync.WaitGroup
-	wg.Add(1)
-	defer wg.Wait()
+func (c *EpinioClient) uploadSources(log logr.Logger, appRef models.AppRef, source string) (string, error) {
+	c.ui.Normal().Msg("Collecting the application sources ...")
+
+	fileInfo, err := os.Stat(source)
+	if err != nil {
+		return "", err
+	}
+
+	archive := source
+
+	if fileInfo.IsDir() {
+		// package directory as archive/tarball
+		tmpDir, tarball, err := helpers.Tar(source)
+		defer os.RemoveAll(tmpDir)
+		if err != nil {
+			return "", err
+		}
+		archive = tarball
+	}
+
+	c.ui.Normal().Msg("Uploading application code ...")
+
+	log.V(1).Info("upload code")
+
+	// open the archive
+	file, err := os.Open(archive)
+	if err != nil {
+		return "", errors.Wrap(err, "failed to open archive")
+	}
+	defer file.Close()
+
+	upload, err := c.API.AppUpload(appRef.Namespace, appRef.Name, file)
+	if err != nil {
+		return "", err
+	}
+	log.V(3).Info("upload response", "response", upload)
+
+	return upload.BlobUID, nil
+}
+
+func (c *EpinioClient) stageLogs(appRef models.AppRef, stageID string) {
 	go func() {
-		defer wg.Done()
-		err := c.AppLogs(appRef.Name, stageID, true, stopChan)
+		err := c.AppLogs(appRef.Name, stageID, true)
 		if err != nil {
 			c.ui.Problem().Msg(fmt.Sprintf("failed to tail logs: %s", err.Error()))
 		}
 	}()
+}
 
-	details.Info("wait for pipelinerun", "StageID", stageID)
-	c.ui.ProgressNote().KeeplineUnder(1).Msg("Running staging")
-
-	_, err := c.API.StagingComplete(appRef.Org, stageID)
-	if err != nil {
-		stopChan <- true // Stop the printing go routine
-		return errors.Wrap(err, "waiting for staging failed")
+func routeMessage(msg *termui.Message, routes []string) *termui.Message {
+	if routes == nil {
+		return msg.WithStringValue("Routes", "<<default>>")
 	}
-	stopChan <- true // Stop the printing go routine
+	if len(routes) == 0 {
+		return msg.WithStringValue("Routes", "<<none>>")
+	}
 
-	return err
+	msg = msg.WithStringValue("Routes", "")
+	sort.Strings(routes)
+	for i, d := range routes {
+		msg = msg.WithStringValue(strconv.Itoa(i+1), d)
+	}
+
+	return msg
+}
+
+func (c *EpinioClient) reportOK(appRef models.AppRef, builder string, routes []string) {
+	msg := c.ui.Success().
+		WithStringValue("Name", appRef.Name).
+		WithStringValue("Namespace", appRef.Namespace).
+		WithStringValue("Builder Image", builder).
+		WithStringValue("Routes", "")
+
+	if len(routes) > 0 {
+		sort.Strings(routes)
+		for i, r := range routes {
+			msg = msg.WithStringValue(strconv.Itoa(i+1), r)
+		}
+	}
+	msg.Msg("App is online.")
 }
