@@ -1,5 +1,15 @@
-// Package auth collects structures and functions around the domains
-// the client works with.
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package domain collects structures and functions around the domains the client works with.
 package domain
 
 import (
@@ -7,18 +17,18 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/epinio/epinio/deployments"
 	"github.com/epinio/epinio/helpers/kubernetes"
+	"github.com/epinio/epinio/internal/helmchart"
 	"github.com/pkg/errors"
 )
 
-// mainDomain is the memoization cache for the name of the main domain
-// of the currently accessed epinio installation.
+// mainDomain is the memoization cache for the name of the main domain of the currently accessed
+// epinio installation.
 var mainDomain = ""
 
-// AppDefaultRoute constructs and returns an application's default
-// route from the main domain and the name of the application
-func AppDefaultRoute(ctx context.Context, name string) (string, error) {
+// AppDefaultRoute constructs and returns an application's default route constructed from the main
+// domain, and the name of the application.
+func AppDefaultRoute(ctx context.Context, name, namespace string) (string, error) {
 	mainDomain, err := MainDomain(ctx)
 	if err != nil {
 		return "", err
@@ -26,12 +36,11 @@ func AppDefaultRoute(ctx context.Context, name string) (string, error) {
 	return fmt.Sprintf("%s.%s", name, mainDomain), nil
 }
 
-// MainDomain determines the name of the main domain of the currently
-// accessed epinio installation. The result is cached in-memory (see
-// variable mainDomain). The function preferably returns cached data,
-// and queries the cluster ingresses only the first time the data is
-// asked for. This is especially useful for long running commands. In
-// other other words, epinio's API server.
+// MainDomain determines the name of the main domain of the currently accessed epinio
+// installation. The result is cached in-memory (see variable mainDomain). The function preferably
+// returns cached data, and queries the cluster ingresses only the first time the data is asked
+// for. This is especially useful for long running commands. In other other words, epinio's API
+// server.
 func MainDomain(ctx context.Context) (string, error) {
 	if mainDomain != "" {
 		return mainDomain, nil
@@ -43,7 +52,7 @@ func MainDomain(ctx context.Context) (string, error) {
 	}
 
 	// Get the epinio ingress
-	ingresses, err := cluster.ListIngress(ctx, deployments.EpinioDeploymentID, "app.kubernetes.io/name=epinio")
+	ingresses, err := cluster.ListIngress(ctx, helmchart.Namespace(), "app.kubernetes.io/name=epinio")
 	if err != nil {
 		return "", errors.Wrap(err, "failed to list ingresses for epinio")
 	}
