@@ -1,8 +1,22 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 // Package duration defines the various durations used throughout
 // Epinio, as timeouts, and other.
 package duration
 
 import (
+	"log"
+	"os"
+	"strconv"
 	"time"
 
 	flag "github.com/spf13/pflag"
@@ -10,24 +24,15 @@ import (
 )
 
 const (
-	systemDomain        = 2 * time.Minute
-	appReady            = 2 * time.Minute
-	deployment          = 10 * time.Minute
-	orgDeletion         = 5 * time.Minute
-	serviceSecret       = 5 * time.Minute
-	serviceProvision    = 5 * time.Minute
-	serviceLoadBalancer = 5 * time.Minute
-	podReady            = 5 * time.Minute
+	deployment          = 3 * time.Minute
+	namespaceDeletion   = 5 * time.Minute
+	configurationSecret = 5 * time.Minute
 	appBuilt            = 10 * time.Minute
-	warmupJobReady      = 30 * time.Minute
-	certManagerReady    = 5 * time.Minute
-	kubedReady          = 5 * time.Minute
 	secretCopied        = 5 * time.Minute
 
 	// Fixed. __Not__ affected by the multiplier.
-	pollInterval = 3 * time.Second
-	userAbort    = 5 * time.Second
-	logHistory   = 48 * time.Hour
+	userAbort  = 5 * time.Second
+	logHistory = 48 * time.Hour
 
 	// Fixed. Standard number of attempts to retry various operations.
 	RetryMax = 10
@@ -36,25 +41,16 @@ const (
 // Flags adds to viper flags
 func Flags(pf *flag.FlagSet, argToEnv map[string]string) {
 	pf.IntP("timeout-multiplier", "", 1, "Multiply timeouts by this factor")
-	viper.BindPFlag("timeout-multiplier", pf.Lookup("timeout-multiplier"))
+	err := viper.BindPFlag("timeout-multiplier", pf.Lookup("timeout-multiplier"))
+	if err != nil {
+		log.Fatal(err)
+	}
 	argToEnv["timeout-multiplier"] = "EPINIO_TIMEOUT_MULTIPLIER"
 }
 
 // Multiplier returns the currently active timeout multiplier value
 func Multiplier() time.Duration {
 	return time.Duration(viper.GetInt("timeout-multiplier"))
-}
-
-// ToCertManagerReady returns the duration to wait until giving up on
-// the cert manager deployment to become ready.
-func ToCertManagerReady() time.Duration {
-	return Multiplier() * certManagerReady
-}
-
-// ToKubedReady returns the duration to wait until giving up on the
-// kube demon deployment to become ready.
-func ToKubedReady() time.Duration {
-	return Multiplier() * kubedReady
 }
 
 // ToSecretCopied returns the duration to wait until giving up on a
@@ -69,68 +65,26 @@ func ToAppBuilt() time.Duration {
 	return Multiplier() * appBuilt
 }
 
-// ToPodReady returns the duration to wait until giving up on getting
-// a system domain
-func ToPodReady() time.Duration {
-	return Multiplier() * podReady
-}
-
-// ToWarmupJobReady return the duration to wait until the builder image
-// warm up job is Complete. The time it takes for that Job to complete depends
-// on the network speed of the cluster so be generous here.
-func ToWarmupJobReady() time.Duration {
-	return Multiplier() * warmupJobReady
-}
-
-// ToSystemDomain returns the duration to wait until giving on getting
-// the system domain
-func ToSystemDomain() time.Duration {
-	return Multiplier() * systemDomain
-}
-
-// ToAppReady returns the duration to wait until the curl request
-// on app url
-func ToAppReady() time.Duration {
-	return Multiplier() * appReady
-}
-
 // ToDeployment returns the duration to wait for parts of a deployment
 // to become ready
 func ToDeployment() time.Duration {
 	return Multiplier() * deployment
 }
 
-// ToOrgDeletion returns the duration to wait for deletion of namespace
-func ToOrgDeletion() time.Duration {
-	return Multiplier() * orgDeletion
+// ToNamespaceDeletion returns the duration to wait for deletion of namespace
+func ToNamespaceDeletion() time.Duration {
+	return Multiplier() * namespaceDeletion
 }
 
-// ToServiceSecret returns the duration to wait for the secret to a
-// catalog service binding to appear
-func ToServiceSecret() time.Duration {
-	return Multiplier() * serviceSecret
-}
-
-// ToServiceProvision returns the duration to wait for a catalog
-// service instance to be provisioned
-func ToServiceProvision() time.Duration {
-	return Multiplier() * serviceProvision
-}
-
-// ToServiceLoadBalancer
-func ToServiceLoadBalancer() time.Duration {
-	return Multiplier() * serviceLoadBalancer
+// ToConfigurationSecret returns the duration to wait for the secret to a
+// catalog configuration binding to appear
+func ToConfigurationSecret() time.Duration {
+	return Multiplier() * configurationSecret
 }
 
 //
 // The following durations are not affected by the timeout multiplier.
 //
-
-// PollInterval returns the duration to use between polls of some kind
-// of check.
-func PollInterval() time.Duration {
-	return pollInterval
-}
 
 // UserAbort returns the duration to wait when the user is given the
 // chance to abort an operation
@@ -139,6 +93,12 @@ func UserAbort() time.Duration {
 }
 
 // LogHistory returns the duration to reach into the past for tailing logs.
+// LogHistory returns the duration to reach into the past for tailing logs.
 func LogHistory() time.Duration {
+	if hours := os.Getenv("LOG_HISTORY_HOURS"); hours != "" {
+		if h, err := strconv.ParseInt(hours, 10, 64); err == nil {
+			return time.Duration(h) * time.Hour
+		}
+	}
 	return logHistory
 }
