@@ -1,20 +1,32 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 // Package s3manager implements the various functions needed to store and retrieve
 // files from an S3 API compatible endpoint (AWS S3, Minio, etc)
 package s3manager
 
 import (
 	"context"
-	"fmt"
-	"strconv"
+	"crypto/x509"
+	"io"
+	"net/http"
 
 	"github.com/epinio/epinio/helpers/kubernetes"
+	"github.com/epinio/epinio/internal/helmchart"
 	"github.com/google/uuid"
 	minio "github.com/minio/minio-go/v7"
 	"github.com/minio/minio-go/v7/pkg/credentials"
 	"github.com/pkg/errors"
+	"github.com/spf13/viper"
 	"gopkg.in/ini.v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type Manager struct {
@@ -23,27 +35,13 @@ type Manager struct {
 }
 
 type ConnectionDetails struct {
-	Endpoint string
-	// UseSSL toggles the use of SSL for the s3 connection. If we deploy
-	// our own thing, linkerd will take care of this and we set UseSSL to
-	// false.
-	// External S3 implementations should use SSL.
+	Endpoint        string
 	UseSSL          bool
 	AccessKeyID     string
 	SecretAccessKey string
 	Bucket          string
 	Location        string
-}
-
-func NewConnectionDetails(endpoint, key, secret, bucket, location string, useSSL bool) *ConnectionDetails {
-	return &ConnectionDetails{
-		Endpoint:        endpoint,
-		UseSSL:          useSSL,
-		AccessKeyID:     key,
-		SecretAccessKey: secret,
-		Bucket:          bucket,
-		Location:        location,
-	}
+	CA              []byte
 }
 
 // Validate makes sure the provided S3 settings are valid
@@ -76,15 +74,36 @@ func (details *ConnectionDetails) Validate() error {
 
 // New returns an instance of an s3 manager
 func New(connectionDetails ConnectionDetails) (*Manager, error) {
-	useSSL := connectionDetails.UseSSL
+	transport := http.DefaultTransport.(*http.Transport).Clone()
+
+	opts := &minio.Options{
+		Creds:     credentials.NewStaticV4(connectionDetails.AccessKeyID, connectionDetails.SecretAccessKey, ""),
+		Secure:    connectionDetails.UseSSL,
+		Transport: transport,
+		Region:    connectionDetails.Location,
+	}
+
+	// if no credentials are provided then we are going to try to connect with the IAM Role
+	if connectionDetails.AccessKeyID == "" && connectionDetails.SecretAccessKey == "" {
+		opts.Creds = credentials.NewIAM("")
+	}
+
+	if len(connectionDetails.CA) > 0 {
+		rootCAs := x509.NewCertPool()
+		if ok := rootCAs.AppendCertsFromPEM(connectionDetails.CA); !ok {
+			return nil, errors.New("cannot append minio ca from connection details to client")
+		}
+
+		tlsConfig := transport.TLSClientConfig.Clone()
+		tlsConfig.RootCAs = rootCAs
+
+		opts.Transport.(*http.Transport).TLSClientConfig = tlsConfig
+	}
 
 	minioClient, err := minio.New(
 		connectionDetails.Endpoint,
-		&minio.Options{
-			Creds:  credentials.NewStaticV4(connectionDetails.AccessKeyID, connectionDetails.SecretAccessKey, ""),
-			Secure: useSSL,
-			Region: connectionDetails.Location,
-		})
+		opts,
+	)
 	if err != nil {
 		return nil, err
 	}
@@ -99,7 +118,7 @@ func New(connectionDetails ConnectionDetails) (*Manager, error) {
 
 // GetConnectionDetails retrieves s3 details from an ini file stored in a
 // secret.
-// Note: The CLI tool in the pipeline task will use the ini file directly.
+// Note: The CLI tool in the staging job will use the ini file directly.
 func GetConnectionDetails(ctx context.Context, cluster *kubernetes.Cluster, secretNamespace, secretName string) (ConnectionDetails, error) {
 	details := ConnectionDetails{}
 	secret, err := cluster.GetSecret(ctx, secretNamespace, secretName)
@@ -125,37 +144,54 @@ func GetConnectionDetails(ctx context.Context, cluster *kubernetes.Cluster, secr
 	}
 	details.Bucket = string(secret.Data["bucket"])
 
+	// load the certificate for S3 if defined
+	s3CertificateSecret := viper.GetString("s3-certificate-secret")
+	if s3CertificateSecret != "" {
+		secret, err = cluster.GetSecret(ctx, helmchart.Namespace(), s3CertificateSecret)
+		if err != nil {
+			return details, errors.Wrapf(err, "getting s3 certificate secret %s", s3CertificateSecret)
+		}
+
+		details.CA = secret.Data["tls.crt"]
+		if ca, ok := secret.Data["ca.crt"]; ok {
+			details.CA = append(details.CA, ca...)
+		}
+	}
+
 	return details, nil
 }
 
-// StoreConnectionDetails stores the S3 connection details in a secret. The
-// ini-file format is compatible with awscli.
-// related tekton task: https://hub.tekton.dev/tekton/task/aws-cli
-func StoreConnectionDetails(ctx context.Context, cluster *kubernetes.Cluster, secretNamespace, secretName string, details ConnectionDetails) (*corev1.Secret, error) {
-	credentials := fmt.Sprintf(`[default]
-aws_access_key_id     = %s
-aws_secret_access_key = %s
-`, details.AccessKeyID, details.SecretAccessKey)
-	config := fmt.Sprintf(`[default]
-region = %s
-`, details.Location)
+// Meta retrieves the meta data for the blob specified by it blobUID.
+func (m *Manager) Meta(ctx context.Context, blobUID string) (map[string]string, error) {
+	blobInfo, err := m.minioClient.StatObject(ctx, m.connectionDetails.Bucket,
+		blobUID, minio.StatObjectOptions{})
+	if err != nil {
+		return map[string]string{}, errors.Wrap(err, "reading the object meta data")
+	}
 
-	secret, err := cluster.Kubectl.CoreV1().Secrets(secretNamespace).Create(ctx,
-		&corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: secretName,
-			},
-			StringData: map[string]string{
-				"credentials": credentials,
-				"config":      config,
-				"endpoint":    details.Endpoint,
-				"useSSL":      strconv.FormatBool(details.UseSSL),
-				"bucket":      details.Bucket,
-			},
-			Type: "Opaque",
-		}, metav1.CreateOptions{})
+	return blobInfo.UserMetadata, nil
+}
 
-	return secret, err
+// UploadStream uploads the given Reader to the S3 endpoint and returns a blobUID which
+// can later be used to fetch the same file.
+func (m *Manager) UploadStream(ctx context.Context, file io.Reader, size int64, metadata map[string]string) (string, error) {
+	if err := m.EnsureBucket(ctx); err != nil {
+		return "", errors.Wrap(err, "ensuring bucket")
+	}
+
+	objectName := uuid.New().String()
+	contentType := "application/tar"
+
+	_, err := m.minioClient.PutObject(ctx, m.connectionDetails.Bucket,
+		objectName, file, size, minio.PutObjectOptions{
+			ContentType:  contentType,
+			UserMetadata: metadata,
+		})
+	if err != nil {
+		return "", errors.Wrap(err, "writing the new object")
+	}
+
+	return objectName, nil
 }
 
 // Upload uploads the given file to the S3 endpoint and returns a blobUID which
@@ -169,7 +205,10 @@ func (m *Manager) Upload(ctx context.Context, filepath string, metadata map[stri
 	contentType := "application/tar"
 
 	_, err := m.minioClient.FPutObject(ctx, m.connectionDetails.Bucket,
-		objectName, filepath, minio.PutObjectOptions{ContentType: contentType, UserMetadata: metadata})
+		objectName, filepath, minio.PutObjectOptions{
+			ContentType:  contentType,
+			UserMetadata: metadata,
+		})
 	if err != nil {
 		return "", errors.Wrap(err, "writing the new object")
 	}
@@ -181,7 +220,7 @@ func (m *Manager) Upload(ctx context.Context, filepath string, metadata map[stri
 func (m *Manager) EnsureBucket(ctx context.Context) error {
 	exists, err := m.minioClient.BucketExists(ctx, m.connectionDetails.Bucket)
 	if err != nil {
-		return err
+		return errors.Wrapf(err, "checking bucket %s exists", m.connectionDetails.Bucket)
 	}
 	if exists {
 		return nil
