@@ -1,109 +1,70 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package client
 
 import (
-	"encoding/json"
 	"fmt"
-	"time"
+	"net/url"
 
-	"github.com/avast/retry-go"
-
-	"github.com/epinio/epinio/helpers"
 	api "github.com/epinio/epinio/internal/api/v1"
-	"github.com/epinio/epinio/internal/duration"
 	"github.com/epinio/epinio/pkg/api/core/v1/models"
 )
 
 // NamespaceCreate creates a namespace
-func (c *Client) NamespaceCreate(req models.NamespaceCreateRequest) (models.Response, error) {
-	var resp models.Response
+func (c *Client) NamespaceCreate(request models.NamespaceCreateRequest) (models.Response, error) {
+	response := models.Response{}
+	endpoint := api.Routes.Path("Namespaces")
 
-	b, err := json.Marshal(req)
-	if err != nil {
-		return resp, nil
-	}
-
-	details := c.log.V(1) // NOTE: Increment of level, not absolute.
-
-	var data []byte
-	err = retry.Do(
-		func() error {
-			details.Info("create org", "org", req.Name)
-			data, err = c.post(api.Routes.Path("Namespaces"), string(b))
-			return err
-		},
-		retry.RetryIf(func(err error) bool {
-			if r, ok := err.(interface{ StatusCode() int }); ok {
-				return helpers.RetryableCode(r.StatusCode())
-			}
-			retry := helpers.Retryable(err.Error())
-
-			details.Info("create error", "error", err.Error(), "retry", retry)
-			return retry
-		}),
-		retry.OnRetry(func(n uint, err error) {
-			details.WithValues(
-				"tries", fmt.Sprintf("%d/%d", n, duration.RetryMax),
-				"error", err.Error(),
-			).Info("Retrying to create namespace")
-		}),
-		retry.Delay(time.Second),
-		retry.Attempts(duration.RetryMax),
-	)
-	if err != nil {
-		return resp, err
-	}
-
-	if err := json.Unmarshal(data, &resp); err != nil {
-		return resp, err
-	}
-
-	return resp, nil
+	return Post(c, endpoint, request, response)
 }
 
 // NamespaceDelete deletes a namespace
-func (c *Client) NamespaceDelete(org string) (models.Response, error) {
-	resp := models.Response{}
+func (c *Client) NamespaceDelete(namespaces []string) (models.Response, error) {
+	response := models.Response{}
 
-	data, err := c.delete(api.Routes.Path("NamespaceDelete", org))
-	if err != nil {
-		return resp, err
+	queryParams := url.Values{}
+	for _, namespace := range namespaces {
+		queryParams.Add("namespaces[]", namespace)
 	}
 
-	if err := json.Unmarshal(data, &resp); err != nil {
-		return resp, err
-	}
+	endpoint := fmt.Sprintf(
+		"%s?%s",
+		api.Routes.Path("NamespaceBatchDelete"),
+		queryParams.Encode(),
+	)
 
-	return resp, nil
+	return Delete(c, endpoint, nil, response)
+}
+
+// NamespaceShow shows a namespace
+func (c *Client) NamespaceShow(namespace string) (models.Namespace, error) {
+	response := models.Namespace{}
+	endpoint := api.Routes.Path("NamespaceShow", namespace)
+
+	return Get(c, endpoint, response)
 }
 
 // NamespacesMatch returns all matching namespaces for the prefix
 func (c *Client) NamespacesMatch(prefix string) (models.NamespacesMatchResponse, error) {
-	resp := models.NamespacesMatchResponse{}
+	response := models.NamespacesMatchResponse{}
+	endpoint := api.Routes.Path("NamespacesMatch", prefix)
 
-	data, err := c.get(api.Routes.Path("NamespacesMatch", prefix))
-	if err != nil {
-		return resp, err
-	}
-
-	if err := json.Unmarshal(data, &resp); err != nil {
-		return resp, err
-	}
-
-	return resp, nil
+	return Get(c, endpoint, response)
 }
 
 // Namespaces returns a list of namespaces
 func (c *Client) Namespaces() (models.NamespaceList, error) {
-	resp := models.NamespaceList{}
+	response := models.NamespaceList{}
+	endpoint := api.Routes.Path("Namespaces")
 
-	data, err := c.get(api.Routes.Path("Namespaces"))
-	if err != nil {
-		return resp, err
-	}
-
-	if err := json.Unmarshal(data, &resp); err != nil {
-		return resp, err
-	}
-
-	return resp, nil
+	return Get(c, endpoint, response)
 }
